feat(AreaModelV2): populate `adminVersions` on save
Just focused on gadm 3.6 right now.

GTC-3086

:white_check_mark: test(AreaModel): use in-memory mongodb for fast tests
For the unit and integration tests we can use
a simple in-memory mongodb. This makes the
tests fast and doesn't require a separate
running mongodb instance.

The e2e tests still use a traditional mongodb
instance.

:white_check_mark: test(AreaModelV2): add tests for when creating adminVersions fails
This is required so we can learn about
problematic areas before relying on adminVersions.